+++ tests/unit/lib/PVE/HA/Config.pm
@@ -1,0 +1,9 @@
+package PVE::HA::Config;
+
+use strict;
+use warnings;
+
+sub read_resources_config { die "unexpected HA resources read\n" }
+sub read_manager_status { die "unexpected HA manager-status read\n" }
+
+1;